+++ pkg/selfupdate/replacer_xattr_other.go
@@ -1,0 +1,11 @@
+//go:build !darwin
+
+package selfupdate
+
+func clearUpdateXattrs(_ string) error {
+	return nil
+}
+
+func copyReplacementFile(sourcePath, targetPath string) error {
+	return copyFile(sourcePath, targetPath)
+}